Text justification.
For the last line of text, it should be left justified and no extra space is inserted between words For example, words:: ['This', 'is', 'an', 'example', 'of', 'text', 'justification.'] L: 16. Return the formatted lines as:['This is an', 'example of text', 'justification. ' ] Note: Each word is guaranteed not to exceed L in length. in java