Intersection Point Of Linked List
Two singly linked-lists intersect at a node and merge into a single linked-list. The size of each is unknown and the number of nodes prior to the intersection node can vary in each list. Devise an algorithm to find the intersecting node.In other words, two singly linked-lists meet at some point, find the node at which they meet in java