Create ThrowExcep.java
Given N pairs of integers, write an algorithm to sort the pairs so that the first number in each pair is sorted increasingly while the second number in each pair is sorted decreasingly. The first and second numbers in each pair can be swapped. Sometimes there will be no solution, in that case throw an exception.